adds blog starter page
adds layout for career and blog pages
