+++ src/pages/blog.js
@@ -14,7 +14,9 @@ import { Container, Col, Row, Button, Image, Badge } from "react-bootstrap"
 export default function Blog() {
   return (
     <Layout>
-      <p>I AM BLOG</p>
+      <h1>AWESOME BLOG TITLE</h1>
+      <p>Lorem ipsum dolor sit amet consectetur, adipisicing elit. Exercitationem, at sequi totam ad fugiat dolore non necessitatibus eveniet ipsa illo minima corrupti modi vero! Unde facilis est pariatur quis? Iure!</p>
+      <p>Lorem ipsum, dolor sit amet consectetur adipisicing elit. Enim dicta consequuntur perspiciatis, vero adipisci, laboriosam dolores omnis possimus est officia doloribus quia? Veritatis dolores sapiente perferendis a dolorem autem porro.s</p>
     </Layout>
   )
 }
